requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-Migrate==4.0.5
Flask-Mail==0.9.1
Flask-WTF==1.1.1
Flask-Babel==4.0.0
WTForms==3.0.1
Werkzeug==2.3.7
python-dotenv==1.0.0
psycopg2-binary==2.9.7
gunicorn==21.2.0
Pillow==10.0.1
markdown==3.5.1
bleach==6.1.0
email-validator==2.0.0
itsdangerous==2.1.2
